5.5.4 Creating a custom directive
Listing 5.25. The contents of the counter.directive.ts file in the src/app/store folder

Listing 5.27. Replacing the built-in directive in the store.component.html file in the src/app/store folder

TBD

Listing 5.26. Registering the custom directive in the store.module.ts file in the src/app/store folder